+++ www/gatsby-config.js
@@ -104,13 +104,6 @@ module.exports = {
     {
       resolve: `gatsby-source-filesystem`,
       options: {
-        name: `ecosystem`,
-        path: `${__dirname}/src/data/ecosystem/`,
-      },
-    },
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
         name: `guidelines`,
         path: `${__dirname}/src/data/guidelines/`,
       },
@@ -166,9 +159,9 @@ module.exports = {
           {
             resolve: `gatsby-remark-autolink-headers`,
             options: {
-              offsetY: 104
+              offsetY: 104,
             },
-          },,
+          },
           `gatsby-remark-copy-linked-files`,
           `gatsby-remark-smartypants`,
         ],
@@ -197,7 +190,7 @@ module.exports = {
           {
             resolve: `gatsby-remark-autolink-headers`,
             options: {
-              offsetY: 104
+              offsetY: 104,
             },
           },
           {