+++ www/src/templates/template-docs-markdown.js
@@ -4,14 +4,9 @@ import React from "react"
 import { Helmet } from "react-helmet"
 import { graphql } from "gatsby"
 import { MDXRenderer } from "gatsby-plugin-mdx"
-import { mediaQueries } from "../gatsby-plugin-theme-ui"
+import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 
 import Layout from "../components/layout"
-import {
-  itemListDocs,
-  itemListTutorial,
-  itemListContributing,
-} from "../utils/sidebar/item-list"
 import MarkdownPageFooter from "../components/markdown-page-footer"
 import DocSearchContent from "../components/docsearch-content"
 import TableOfContents from "../components/docs-table-of-contents"
@@ -31,39 +26,29 @@ const containerStyles = {
   px: 9,
 }
 
-const getDocsData = slug => {
-  const [urlSegment] = slug.split(`/`).slice(1)
-  const itemListLookup = {
-    docs: itemListDocs,
-    contributing: itemListContributing,
-    tutorial: itemListTutorial,
-  }
-
-  return [urlSegment, itemListLookup[urlSegment]]
-}
-
 function DocsTemplate({ data, location, pageContext: { next, prev } }) {
   const page = data.mdx
-  const [urlSegment, itemList] = getDocsData(page.fields.slug)
+  const [urlSegment] = page.fields.slug.split(`/`).slice(1)
   const toc =
     !page.frontmatter.disableTableOfContents && page.tableOfContents.items
+
+  const description = page.frontmatter.description || page.excerpt
 
   return (
     <React.Fragment>
       <Helmet>
         <title>{page.frontmatter.title}</title>
-        <meta name="description" content={page.excerpt} />
-        <meta property="og:description" content={page.excerpt} />
+        <meta name="description" content={description} />
+        <meta property="og:description" content={description} />
         <meta property="og:title" content={page.frontmatter.title} />
         <meta property="og:type" content="article" />
-        <meta name="twitter:description" content={page.excerpt} />
+        <meta name="twitter:description" content={description} />
         <meta name="twitter.label1" content="Reading time" />
         <meta name="twitter:data1" content={`${page.timeToRead} min read`} />
       </Helmet>
       <Layout
         location={location}
         locale={page.fields.locale}
-        itemList={itemList}
         enableScrollSync={urlSegment === `docs` ? false : true}
       >
         <DocSearchContent>
@@ -78,7 +63,7 @@ function DocsTemplate({ data, location, pageContext: { next, prev } }) {
               },
             }}
           >
-            <Breadcrumb location={location} itemList={itemList} />
+            <Breadcrumb location={location} />
             <h1 id={page.fields.anchor} sx={{ mt: 0 }}>
               {page.frontmatter.title}
             </h1>
@@ -163,6 +148,7 @@ export const pageQuery = graphql`
       }
       frontmatter {
         title
+        description
         overview
         issue
         disableTableOfContents